+++ src/Api/Providers/News.php
@@ -2,7 +2,6 @@
 
 namespace seregazhuk\PinterestBot\Api\Providers;
 
-use seregazhuk\PinterestBot\Api\Request;
 use seregazhuk\PinterestBot\Helpers\UrlHelper;
 
 class News extends Provider
@@ -14,12 +13,8 @@ class News extends Provider
      *
      * @return array
      */
-    public function latest()
+    public function last()
     {
-        $data = ['options' => ['allow_state' => true]];
-        $query = Request::createQuery($data);
-        $response = $this->request->exec(UrlHelper::RESOURCE_GET_LATEST_NEWS."?{$query}");
-
-        return $this->response->getData($response);
+        return $this->execGetRequest(['allow_state' => true], UrlHelper::RESOURCE_GET_LATEST_NEWS);
     }
 }