+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [3.0.0] - 2016-05-29
+### Changed:
+ - News *latest* method renamed to *last*
+ - Request and providers refactoring
+ 
+### Fixes:
+ - Boards follow/unfollow request
+
 ## [2.6.0] - 2016-05-28
 ### Updated:
  - Pagination refactoring